Module Bit_counter

type t

val sexp_of_t : t -> Sexplib0.Sexp.t

val create : len:int -> t

val blit_next_value : t -> dst:Bit_array.t -> dst_pos:int -> unit

Blit the next value encoded on len bits into the destination array starting from dst_pos. When the internal counter has reached its maximum value, it starts again from 0.