Module Bit_matrix

Extending Bit_array for arrays of 2 dimensions. For examples, loading/saving contents of memories.

type t = Bit_array.t array

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val sexp_of_t : t -> Sexplib0.Sexp.t

val init_matrix_linear : dimx:int -> dimy:int -> f:(int -> 'a) -> 'a array array

Create a new matrix initiated with the results of the function f called from left to right with indexes increasing from 0 to n-1. This loops along the dimx dimension first, and dimy second, which means that all the ys for a given x are consecutive).

val of_bit_array : dimx:int -> dimy:int -> Bit_array.t -> t

Initiate a new matrix by taking the contents from a linear bit array. If the input is too long, only the useful prefix is used. If the input is too short, the rest of the bits is set to false.

val of_text_file : dimx:int -> dimy:int -> path:Fpath.t -> t

See Bit_array.of_text_file.

val to_text_file : t -> path:Fpath.t -> unit

Save it to disk. To make things more readable, write things line by line in the file (along the dimx dimension).

val to_text_channel : t -> Core.Out_channel.t -> unit

Dimensions

val dimx : t -> int

val dimy : t -> int