Module Bopkit

module Arithmetic_expression : sig ... end

module Comment : sig ... end

Pretty-printing comments that are found in bopkit files.

module Comments : sig ... end

module Conditional_expression : sig ... end

module Control_structure : sig ... end

Control structure acts as macro. They're developed statically during the analysis of the input program.

module Eval_error : sig ... end

A type used to manipulate the errors that occur when evaluating parameters, string_with_vars, etc.

module Expand_utils : sig ... end

Some utils while expanded elements of a Netlist.t into elements of an Expanded_netlist.t.

module Expanded_netlist : sig ... end

Expanded netlist is an intermediate representation built during the analysis of a circuit. Constructs such as control structures and imbrications have been developed. This representation is used by different parts of bopkit such as the simulator, bop2c, etc.

module Interval : sig ... end

module Netlist : sig ... end

This file defines the abstract syntax tree of *.bop files. For each file, the parser returns a single value of type t.

module Or_eval_error : sig ... end

module Parameter : sig ... end

Parameters are language construct that allows some flexibility in the definition of blocs, adding parameters to external blocks commands, etc. Example:

module Parameters : sig ... end

Bindings between parameter names and values. The type is a list, with the semantic that values added at the front of the list take precedence over other values present in the rest of the list. This allows for functional variable shadowing, by adding more elements to the front of an existing binding.

module String_with_vars : sig ... end

String with variables %{...} in it. This allows commands to depend on parameters, etc.