Module Bopkit.Expanded_netlist

Expanded netlist is an intermediate representation built during the analysis of a circuit. Constructs such as control structures and imbrications have been developed. This representation is used by different parts of bopkit such as the simulator, bop2c, etc.

type memory = {

loc : Loc.t;

name : Base.string;

memory_kind : Netlist.memory_kind;

address_width : Base.int;

data_width : Base.int;

memory_content : Bit_matrix.t Base.option;

}

type external_method = {

method_name : Base.string;

attributes : Base.string Base.list;

}

type external_block = {

loc : Loc.t;

name : Base.string;

attributes : Base.string Base.list;

init_messages : Base.string Base.list;

methods : external_method Base.list;

command : Base.string;

}

Once the control structures have been developed, all the indexes are computed and are integers.

type index =

| Interval of Base.int * Base.int

| Index of Base.int

val equal_index : index -> index -> Base.bool

val sexp_of_index : index -> Sexplib0.Sexp.t

type variable =

| Signal of {

name : Base.string;

}

| Bus of {

loc : Loc.t;

name : Base.string;

indexes : index Base.list;

}

| Internal of Base.int

type variables = {

expanded : Base.string Base.list;

original_grouping : variable Base.list;

}

type call =

| Block of {

name : Base.string;

}

| External_block of {

name : Base.string;

method_name : Base.string Base.option;

external_arguments : Base.string Base.list;

}

type node = {

loc : Loc.t;

call : call;

inputs : variables;

outputs : variables;

}

Flat nodes where the nested inputs have all been developed.

type block = {

loc : Loc.t;

name : Base.string;

attributes : Base.string Base.list;

inputs : variables;

outputs : variables;

unused_variables : variables;

nodes : node Base.list;

}

Blocks with parameters are produced on demand, with a fresh name for each variation of required parameters. At the moment, the actual name is made of a concatenation of the original name plus additional characters taken from the concrete syntax, such as '', '', etc to make it unique. Example: "name4".

type t = {

paths : Fpath.t Base.list;

rom_memories : Bit_matrix.t Base.array;

memories : memory Base.array;

external_blocks : external_block Base.list;

blocks : block Base.list;

main_block_name : Base.string;

}