Module Bopkit.Interval

type t = {

from : Base.int;

to_ : Base.int;

}

from and to_ are included in the interval. from may be smaller than, equal to, or greater than to_.

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val sexp_of_t : t -> Sexplib0.Sexp.t

val singleton : Base.int -> t

val expand : t -> f:(Base.int -> 'a) -> 'a Base.list

val expand_list : t Base.list -> f:(Base.int -> 'a) -> 'a Base.list Base.list

When multiple indexes are used in sequence, the development goes from right to left, that is:

     a:[2]:[3] --->  a[0][0], a[0][1], a[0][2], a[1][0], a[1][1], a[1][2]