Module Bopkit.Netlist

This file defines the abstract syntax tree of *.bop files. For each file, the parser returns a single value of type t.

type include_file = {

loc : Loc.t;

comments : Comments.t;

include_file_kind : include_file_kind;

}

Including other files (with #include).

and include_file_kind =

| File_path of Fpath.t

| Distribution of {

file : Fpath.t;

file_is_quoted : Base.bool;

}

This encodes the difference between including a file from the standard library, and a file from a user's project.

     #include <stdlib.bop>
     #include "my_file.bop"

val equal_include_file : include_file -> include_file -> Base.bool

val equal_include_file_kind : include_file_kind -> include_file_kind -> Base.bool

val sexp_of_include_file : include_file -> Sexplib0.Sexp.t

val sexp_of_include_file_kind : include_file_kind -> Sexplib0.Sexp.t

type parameter = {

loc : Loc.t;

comments : Comments.t;

name : Base.string;

parameter_value : parameter_value;

}

Defining parameters (with #define).

and parameter_value =

| DefCondInt of Conditional_expression.t * Arithmetic_expression.t * Arithmetic_expression.t

| DefInt of Arithmetic_expression.t

| DefCondString of Conditional_expression.t * Base.string * Base.string

| DefString of Base.string

val equal_parameter : parameter -> parameter -> Base.bool

val equal_parameter_value : parameter_value -> parameter_value -> Base.bool

val sexp_of_parameter : parameter -> Sexplib0.Sexp.t

val sexp_of_parameter_value : parameter_value -> Sexplib0.Sexp.t

type memory_kind =

| ROM

| RAM

The 2 different kinds of memory. ROM is read-only, and RAM allows read/write.

val equal_memory_kind : memory_kind -> memory_kind -> Base.bool

val sexp_of_memory_kind : memory_kind -> Sexplib0.Sexp.t

type memory_content =

| Text of Base.string

| File of Fpath.t

| Zero

The different ways to specify the initial contents of memories.

val equal_memory_content : memory_content -> memory_content -> Base.bool

val sexp_of_memory_content : memory_content -> Sexplib0.Sexp.t

type memory = {

loc : Loc.t;

comments : Comments.t;

name : Base.string;

memory_kind : memory_kind;

address_width : Arithmetic_expression.t;

data_width : Arithmetic_expression.t;

memory_content : memory_content;

}

val equal_memory : memory -> memory -> Base.bool

val sexp_of_memory : memory -> Sexplib0.Sexp.t

type external_block_api_element =

| Init_message of {

loc : Loc.t;

comments : Comments.t;

message : Base.string;

}

| Method of {

loc : Loc.t;

comments : Comments.t;

method_name : Base.string;

method_name_is_quoted : Base.bool;

attributes : Base.string Base.list;

}

and external_block = {

loc : Loc.t;

head_comments : Comments.t;

tail_comments : Comments.t;

name : Base.string;

attributes : Base.string Base.list;

api : external_block_api_element Control_structure.t Base.list;

command : Base.string;

}

val equal_external_block_api_element : external_block_api_element -> external_block_api_element -> Base.bool

val equal_external_block : external_block -> external_block -> Base.bool

val sexp_of_external_block_api_element : external_block_api_element -> Sexplib0.Sexp.t

val sexp_of_external_block : external_block -> Sexplib0.Sexp.t

type functional_argument = {

name : Base.string;

name_is_quoted : Base.bool;

}

Functional argument may be added to parameterized functions. When calling a function with functional parameters in its formal definition, there's the question of whether the argument is a string with vars such as <"name_%{var}"> or whether the call is merely an identifier, such as <ident>. The second field distinguishes the two.

val equal_functional_argument : functional_argument -> functional_argument -> Base.bool

val sexp_of_functional_argument : functional_argument -> Sexplib0.Sexp.t

type index =

| Segment of Arithmetic_expression.t

| Interval of Arithmetic_expression.t * Arithmetic_expression.t

| Index of Arithmetic_expression.t

Variables may be followed by indexes, to expand signals into buses, or collections of buses. Here are example of indexes, along with their expanded version.

      a[0]     --->  a[0]
      a[0..3]  --->  a[0], a[1], a[2], a[3]
      a:[4]    --->  a[0], a[1], a[2], a[3]
      a:[-4]   --->  a[3], a[2], a[1], a[0]

When multiple indexes are used in sequence, the development goes from right to left, that is:

      a:[2]:[3] --->  a[0][0], a[0][1], a[0][2], a[1][0], a[1][1], a[1][2]

val equal_index : index -> index -> Base.bool

val sexp_of_index : index -> Sexplib0.Sexp.t

type variable =

| Signal of {

name : Base.string;

}

| Bus of {

loc : Loc.t;

name : Base.string;

indexes : index Base.list;

}

val equal_variable : variable -> variable -> Base.bool

val sexp_of_variable : variable -> Sexplib0.Sexp.t

type external_call_output_size =

| Inferred

| Specified of Arithmetic_expression.t

In the case of a call to an external block, the output size may not always be known. Sometimes it can be inferred, or computed from an expression.

val equal_external_call_output_size : external_call_output_size -> external_call_output_size -> Base.bool

val sexp_of_external_call_output_size : external_call_output_size -> Sexplib0.Sexp.t

type call =

| Block of {

name : Base.string;

arguments : Arithmetic_expression.t Base.list;

functional_arguments : functional_argument Base.list;

}

| External_block of {

name : Base.string;

method_name : Base.string Base.option;

method_name_is_quoted : Base.bool;

external_arguments : Base.string Base.list;

output_size : external_call_output_size;

}

| External_command of {

command : Base.string;

output_size : external_call_output_size;

}

The different kinds of calls to another block. The block that is called can be a block previously defined, and an external block, or a external construct. The syntax is different for all three:

      output = block(input);
      output = $block.method(input);
      output = external("./command.exe", input);

val equal_call : call -> call -> Base.bool

val sexp_of_call : call -> Sexplib0.Sexp.t

type nested_inputs =

| Variables of {

loc : Loc.t;

comments : Comments.t;

variables : variable Base.list;

}

| Nested_node of {

loc : Loc.t;

comments : Comments.t;

call : call;

inputs : nested_inputs Base.list;

}

Node's input may be variables or the outputs of calls to nested nodes.

Here is an example of a node with nested inputs:

      e = or(and(a, b), xor(c, d));

val equal_nested_inputs : nested_inputs -> nested_inputs -> Base.bool

val sexp_of_nested_inputs : nested_inputs -> Sexplib0.Sexp.t

type node = {

loc : Loc.t;

comments : Comments.t;

call : call;

inputs : nested_inputs Base.list;

outputs : variable Base.list;

}

val equal_node : node -> node -> Base.bool

val sexp_of_node : node -> Sexplib0.Sexp.t

module Interface : sig ... end

type block = {

loc : Loc.t;

head_comments : Comments.t;

tail_comments : Comments.t;

name : Interface.t;

attributes : Base.string Base.list;

inputs : variable Base.list;

outputs : variable Base.list;

unused_variables : variable Base.list;

nodes : node Control_structure.t Base.list;

}

The definition of a new block.

val equal_block : block -> block -> Base.bool

val sexp_of_block : block -> Sexplib0.Sexp.t

type t = {

include_files : include_file Base.list;

parameters : parameter Base.list;

memories : memory Base.list;

external_blocks : external_block Base.list;

blocks : block Base.list;

eof_comments : Comments.t;

}

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val sexp_of_t : t -> Sexplib0.Sexp.t