Module Bopkit.Parameter

Parameters are language construct that allows some flexibility in the definition of blocs, adding parameters to external blocks commands, etc. Example:

#define N 4

Bloc_not[N] (a:[N]) = s:[N]
where
  for i = 0 to N-1
    s[i] = not(a[i]);
  end for;
end where;

Parameters can be bound to integers or strings.

module Value : sig ... end

type t = {

name : Base.string;

value : Value.t;

}

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val sexp_of_t : t -> Sexplib0.Sexp.t

val arg_type : t Command.Arg_type.t

Defining a syntax with which parameters may be passed via the command line.