Module Bopkit.Parameters

Bindings between parameter names and values. The type is a list, with the semantic that values added at the front of the list take precedence over other values present in the rest of the list. This allows for functional variable shadowing, by adding more elements to the front of an existing binding.

type t = Parameter.t Base.list

val find : t -> parameter_name:Base.string -> Parameter.Value.t Base.option

val mem : t -> parameter_name:Base.string -> Base.bool

val keys : t -> Base.string Base.list

val overrides : t Command.Param.t

Defining command line parameters that can be used to supply overriding values for parameters contained in a circuit.