Module String_with_vars.Syntax

type t =

| Dollar

| Percent

We are in the process of transitioning from $(...) to %{...}. Meanwhile both syntax are accepted, and the printer can decide which to use.