Module Bopkit_bdd.Block

type t

val sexp_of_t : t -> Sexplib0.Sexp.t

val pp : Stdlib.Format.formatter -> t -> Base.unit

Output a bloc using Bopkit concrete syntax. The resulting simulation of the Bopkit block is equal to the boolean function that the block encodes.

val of_muxtrees : ?block_name:Base.string -> Muxtree.t Base.list -> input_size:Base.int -> t

Create a bloc from a computation where each output is computed by a dedicated muxtree.

val of_muxlist : ?block_name:Base.string -> Muxlist.t -> input_size:Base.int -> t

Create a bloc from a computation where internal tree nodes have been shared.

val number_of_gates : t -> Base.int

Count all gates that cannot be immediately simplified (such as id nodes).