Module Bopkit_bdd.Ident

In bopkit bdd synthesized blocks, there are 3 different classes of signals:

-the input -the output -internal signals used for sharing intermediate results.

The code that prints out the block is free to use any representation it chooses for them, as long as they stay distinct.

type t =

| Input of Base.int

| Output of Base.int

| Internal of Base.int

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : t Base__Ppx_hash_lib.hash_fold

val hash : t -> Base__Ppx_hash_lib.Std.Hash.hash_value

val sexp_of_t : t -> Sexplib0.Sexp.t