Module Bopkit_bdd.Muxtree

type t =

| Constant of Base.bool Base.option

| Signal of Ident.t

| Not_signal of Ident.t

| Mux of {

input : Base.int;

vdd : t;

gnd : t;

}

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : t Base__Ppx_hash_lib.hash_fold

val hash : t -> Base__Ppx_hash_lib.Std.Hash.hash_value

val sexp_of_t : t -> Sexplib0.Sexp.t

include Base.Comparator.S with type t := t

type comparator_witness

val comparator : (t, comparator_witness) Base__Comparator.comparator

val normalize : t -> t

Operate immediate simplifications on t if able.

val of_partial_bit_matrix : Partial_bit_matrix.t -> t Base.list

Construct the non optimized trees from a boolean function f encoded as a partial bit matrix. The returned list r has the size of the words returned by f, with r.(i) being the i-th bit of the output.

val number_of_gates : t -> Base.int