Module Bopkit_circuit

module Cds : sig ... end

According to ChatGPT :

module Circuit : sig ... end

A bopkit circuit holds a Cds.t along with information required to fire the computation of gates that are not entirely combinatorial, such as memories and external nodes.

module Gate : sig ... end

module Gate_kind : sig ... end

module Output_wire : sig ... end