Module Bopkit_circuit.Gate_kind

type t =

| Input

| Output

| Id

| Not

| And

| Or

| Xor

| Mux

| Rom of {

loc : Loc.t;

name : string;

index : int;

}

| Ram of {

loc : Loc.t;

name : string;

address_width : int;

data_width : int;

contents : Bit_matrix.t;

}

| Reg of {

initial_value : bool;

}

| Regr of {

index_of_regt : int;

}

| Regt

| Clock

| Gnd

| Vdd

| External of {

loc : Loc.t;

name : string;

method_name : string option;

arguments : string list;

protocol_prefix : string Core.Set_once.t;
(*
The method name and the arguments are constant for a given gate, thus the simulator caches the prefix of the string that it will send to the external process.
*)

index : int Core.Set_once.t;
(*
The index of the external gate in the process table is determined at runtime during the initialization of the simulation.
*)

}

val sexp_of_t : t -> Sexplib0.Sexp.t

val pp_debug : t -> _ Pp.t

A pretty printer used for debugging. Does not need to (and doesn't) produce concrete syntax.

module Primitive : sig ... end