Module Bopkit_circuit.Output_wire

type t = {

gate_index : int;

input_index : int;

}

Each of the output bit of a gate can be connected to the input of other gates. Since the cds is sorted topologically, such input gate necessarily appears after the output gate in the cds.

An output wire designates exactly which input of which gate the output bit is connected to. Note that an output may be connected to several inputs, so each output may have several output wires. An the other hand, an input may be connected to at most one single output wire.

The gate_index is the array index of the input gate this output wire is connected to in the cds.

The input_index is the array index of the actual input bit of that gate the wire is connected to.

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

include Ppx_hash_lib.Hashable.S with type t := t

val hash_fold_t : t Base__Ppx_hash_lib.hash_fold

val hash : t -> Base__Ppx_hash_lib.Std.Hash.hash_value

val sexp_of_t : t -> Sexplib0.Sexp.t

module Short_sexp : sig ... end