Module Bopkit_compiler

module Expanded_block : sig ... end

This is an internal representation used as an intermediate step in the chain of transformations that leads to the final cds.

module Primitive : sig ... end

Primitives are the building block of circuits. During the various transformations performed as we build the circuit, we refer to an environment of primitives indexed by their name to resolve the calls, and perform arity checks.

module Config : sig ... end

val create_block : Bopkit.Expanded_netlist.block -> error_log:Error_log.t -> primitives:Primitive.env -> env:Expanded_block.env -> Expanded_block.t

val expand_netlist : path:Fpath.t -> error_log:Error_log.t -> config:Config.t -> (Primitive.env * Bopkit.Expanded_netlist.t) Core.Or_error.t

val circuit_of_netlist : path:Fpath.t -> error_log:Error_log.t -> config:Config.t -> Bopkit_circuit.Circuit.t Core.Or_error.t