Module Bopkit_compiler.Expanded_block

This is an internal representation used as an intermediate step in the chain of transformations that leads to the final cds.

An Expanded_block.t is quite close to the blocks found in an Expanded_netlist.t, except we've distinguished the primitives from other calls, and we've done more context verifications, such as arity checks.

The environment formed by all Expanded_block.t indexed by their name is then used to create the Expanded_nodes.t structure.

type call =

| Block of {

name : string;

}

| Primitive of {

gate_kind : Bopkit_circuit.Gate_kind.t;

}

val sexp_of_call : call -> Sexplib0.Sexp.t

type node = {

call : call;

inputs : string list;

outputs : string list;

}

val sexp_of_node : node -> Sexplib0.Sexp.t

type t = {

loc : Loc.t;

name : string;

local_variables : string list;

input_names : string array;

output_names : string array;

nodes : node list;

}

val sexp_of_t : t -> Sexplib0.Sexp.t

type env = t Core.Map.M(Core.String).t

An environment of expanded_blocks, indexed by their name. This allows for performing the lookups when encountering a call equal Block { name } in the body of another block.