Module Bopkit_process_interpreter

Interpreter for a bopkit process program.

val run_program : error_log:Error_log.t -> architecture:Base.int -> program:Bopkit_process.Program.t -> Base.unit Base.Or_error.t

Given an architecture parameter and an input program, interpret its evaluation - reading inputs from stdin and writing output to stdout in a loop until stdin is closed.