Module Bopkit_topological_sort

Topological sort, e.g. used to reorder parameters, functions, etc.

Nodes are compared using hashable keys. sort only keeps the last version of each node (previous versions are discarded), and then order them using a stable topological sort. Node.parents is only called once per node.

module type Node = sig ... end

module type Key = sig ... end

val sort : (module Node with type key = 'key and type t = 'node) -> (module Key with type t = 'key) -> 'node Base.list -> error_log:Error_log.t -> 'node Base.list

Best effort to make parents appear before their children, and otherwise stable. If there is a cycle, sort doesn't raise (but e.g. the evaluation of bopkit parameters is going to find free variables later).