Module Partial_bit

type t = bool option

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

include Ppx_quickcheck_runtime.Quickcheckable.S with type t := t

val quickcheck_generator : t Base_quickcheck.Generator.t

val quickcheck_observer : t Base_quickcheck.Observer.t

val quickcheck_shrinker : t Base_quickcheck.Shrinker.t

val sexp_of_t : t -> Sexplib0.Sexp.t

val conflicts : t -> with_:bool -> bool

Check whether the partial bit agrees with a fully specified value. If the bit is specified, it will expect the given boolean to be equal to its specification. If the bit is unspecified, there is no conflict. The function returns true in case of conflicts.