Module Partial_bit_array

Sometimes some of the bits of a given computation are not specified (sometimes called "outputs with bit don't care"). This modules allows to represent such cases.

In contexts where bits don't care are encoded as human readable characters, we use the star char '*' to designate an unspecified bit.

type t = Partial_bit.t array

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

include Ppx_quickcheck_runtime.Quickcheckable.S with type t := t

val quickcheck_generator : t Base_quickcheck.Generator.t

val quickcheck_observer : t Base_quickcheck.Observer.t

val quickcheck_shrinker : t Base_quickcheck.Shrinker.t

val sexp_of_t : t -> Sexplib0.Sexp.t

val of_01star_chars_in_string : string -> t

Builds a t by only looking at the characters '0', '1' and '*' from a string, and ignores all the other characters.

val to_string : t -> string

Creates a string made of '0', '1' and '*'.

val of_text_file : path:Fpath.t -> t

See Bit_array.of_text_file.

val to_text_file : t -> path:Fpath.t -> unit

Save to disk, ends with a newline.

val to_text_channel : t -> Core.Out_channel.t -> unit

val conflicts : t -> with_:Bit_array.t -> bool

Check whether a partial specification has conflicts with a fully specified value. Unspecified bits can take whatever value in the specified value. If the values do not have the same size, only compare up to the smallest of the two.