Module Partial_bit_matrix

Extending Partial_bit_array for arrays of 2 dimensions.

type t = Partial_bit_array.t array

include Ppx_compare_lib.Comparable.S with type t := t

val compare : t Base__Ppx_compare_lib.compare

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val sexp_of_t : t -> Sexplib0.Sexp.t

val of_partial_bit_array : dimx:int -> dimy:int -> Partial_bit_array.t -> t

Initiate a new matrix by taking the contents from a linear partial bit array. If the input is too long, only the useful prefix is used. If the input is too short, the rest of the bits is set to None.

val of_text_file : dimx:int -> dimy:int -> path:Fpath.t -> t

See Partial_bit_array.of_text_file.

val to_text_file : t -> path:Fpath.t -> unit

val to_text_channel : t -> Core.Out_channel.t -> unit