Module Visa

module Address : sig ... end

module Assembly_instruction : sig ... end

module Constant_name : sig ... end

module Executable : sig ... end

In contrast to a Program.t which may include higher level language constructs such as constants and macros definitions, an executable is a mere sequence of instructions. This is the expanded version of the program once the macros have been resolved.

module Instruction : sig ... end

module Instruction_name : sig ... end

module Label : sig ... end

module Machine_code : sig ... end

module Macro_name : sig ... end

module Parameter_name : sig ... end

module Program : sig ... end

module Register_name : sig ... end