Module Visa.Executable

In contrast to a Program.t which may include higher level language constructs such as constants and macros definitions, an executable is a mere sequence of instructions. This is the expanded version of the program once the macros have been resolved.

There are three flavors of executable that the assembler works with:

with instruction pointers t: the instructions involving jumping to labels referring directly to raw address in the executable code.

with labels With_labels.t: the instructions involving jumping to labels still use named labels as opposed to addresses in the code.

binary: the instruction have been translated from their human readable form into binary codes directly executable by the hardware machine implementing the microprocessor.

module With_labels : sig ... end

module Instruction_pointer : sig ... end

type t = Instruction_pointer.t Instruction.t array

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val sexp_of_t : t -> Sexplib0.Sexp.t

val resolve_labels : With_labels.t -> t

val disassemble : t -> Program.t

module Machine_code : sig ... end

val to_machine_code : t -> Machine_code.t