Module Executable.Instruction_pointer

type t

An instruction pointer is the address of an instruction in the executable. That is the index in the executable array where this instruction is located. It's an integer. In this representation, there is no special introduction of labels anymore, any code location may serve as a label to jump to.

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val sexp_of_t : t -> Sexplib0.Sexp.t

val to_int : t -> int