Module Visa.Machine_code

module Byte : sig ... end

Visa's machine instructions may be encoded on 1 or 2 byte.

type t = Byte.t array

include Ppx_compare_lib.Equal.S with type t := t

val equal : t Base__Ppx_compare_lib.equal

val sexp_of_t : t -> Sexplib0.Sexp.t

val of_text_file_exn : path:Fpath.t -> error_log:Error_log.t -> t

val of_instructions : int Instruction.t array -> t

Compile down instructions into a binary encoding that can be read by the microprocessor.

val to_instructions : t -> path:Fpath.t -> error_log:Error_log.t -> int Instruction.t array

Disassemble the binary code to produce back an original sequence of instructions.

module For_testing : sig ... end