Module Visa_assembler.Assembly_construct

type t = private

| Label_introduction of {

label : Visa.Label.t With_loc.t;

}

| Assembly_instruction of {

assembly_instruction : Visa.Assembly_instruction.t;

}

After the environment is built, the assembler only keep these top level constructs from the original program.