Module Visa_assembler.Environment

type t = private {

constants : Visa.Program.Constant_kind.t With_loc.t Core.Map.M(Visa.Constant_name).t;

macros : Macro_definition.t Core.Map.M(Visa.Macro_name).t;

labels : Visa.Label.t With_loc.t Core.Map.M(Visa.Label).t;

}

All the definitions and labels of a program are gathered in an initial pass to build the environment.

val sexp_of_t : t -> Sexplib0.Sexp.t