Module Visa_dsl

Visa_dsl is an OCaml API to create visa-assembly program. Another way to describe it is that it is an embedded DSL for visa, where the host language is OCaml.

module Session : sig ... end

val program : (Session.t -> unit) -> Visa.Program.t

module O : sig ... end

Advanced uses

module Advanced : sig ... end