Module Visa_simulator

module Code : sig ... end

module Execution_stack : sig ... end

module Memory : sig ... end

t is the part of the internal state of the simulator that deals with values that are stored in ram as well as registers. It does not contain the code pointer.

module Config : sig ... end

It is possible to simulate the execution of a visa program, for example to debug an assembly code, or to quickly compute its resulting execution.

type t = private {

error_log : Error_log.t;

environment : Visa_assembler.Environment.t;

code : Code.t;

execution_stack : Execution_stack.t;

memory : Memory.t;

config : Config.t;

}

val sexp_of_t : t -> Sexplib0.Sexp.t

val create : config:Config.t -> error_log:Error_log.t -> program:Visa.Program.t -> t Core.Or_error.t

Create a new simulator from an assembly program. Having the assembly program is the most useful entry point when debugging, as the simulator will allow for following along with macro applications, etc.

You may simulate the execution of an executable by disassembling it first (see Visa.Executable.disassemble).

module Step_result : sig ... end

val step : t -> error_log:Error_log.t -> Step_result.t Core.Or_error.t

val run : t -> error_log:Error_log.t -> unit Core.Or_error.t

val main : Core.Command.t