Module Visa_simulator.Memory

t is the part of the internal state of the simulator that deals with values that are stored in ram as well as registers. It does not contain the code pointer.

type t

val sexp_of_t : t -> Sexplib0.Sexp.t

val create : unit -> t

val load_initial_memory : t -> Bit_matrix.t -> unit

Blit the memory with a contents read from a file at initialization time.

val output_device : t -> Visa_simulator__.Output_device.t

Get the output device from the machine. The output device is maintained up to date. The caller commits not to side-effect the returned structure, the behavior of the simulation is unspecified otherwise.

val register_value : t -> register_name:Visa.Register_name.t -> int

t contains mutable values for each of the registers. This function allows access to the current contents.

val load : t -> address:Visa.Address.t -> register_name:Visa.Register_name.t -> unit

load t address register_name fetches data from memory at given address and write it to the register whose name is supplied.

val load_value : t -> value:int -> register_name:Visa.Register_name.t -> unit

As opposed to load which involves a fetch to memory, load_value sets the contents of the register whose name is supplied with the given immediate value.

val store : t -> register_name:Visa.Register_name.t -> address:Visa.Address.t -> unit Core.Or_error.t

store t ~register_name ~address is the opposite operation of load - it will write to memory at the given address the value currently contained in the register whose name is supplied.

val write : t -> register_name:Visa.Register_name.t -> address:Visa.Address.t -> unit Core.Or_error.t

write t ~register_name ~address is like store but writes to the output device. It will write to the output device at the given address the value currently contained in the register whose name is supplied.

Operations on registers

Unless specified otherwise, when the operation is a binary operation involving both registers, the result is stored into R1. Example: add mutates R1 with the result of R0 + R1.

Unary operations are available for both registers, as specified by the supplied register_name. Example: not R0 negates the contents of register R0.

val add : t -> unit

val and_ : t -> unit

val switch : t -> unit

val cmp : t -> unit

R1 <- (R0 = R1).

val not_ : t -> register_name:Visa.Register_name.t -> unit