bopkit index

Library bopkit.bdd

The entry point of this library is the module: Bopkit_bdd.

Library bopkit.bdd-command

The entry point of this library is the module: Bopkit_bdd_command.

Library bopkit.bit-utils

This library exposes the following toplevel modules:

Bit_array In bopkit, bool arrays have many uses, including saving or reading the contents of memory-units from files, exchanging the value of buses between external blocks, printing block outputs to stdout, etc.

Bit_counter

Bit_matrix Extending Bit_array for arrays of 2 dimensions. For examples, loading/saving contents of memories.

Bit_string_encoding Some bit encoding that are used throughout the project.

Partial_bit

Partial_bit_array Sometimes some of the bits of a given computation are not specified (sometimes called "outputs with bit don't care"). This modules allows to represent such cases.

Partial_bit_matrix Extending Partial_bit_array for arrays of 2 dimensions.

Library bopkit.block

The entry point of this library is the module: Bopkit_block.

Library bopkit.bopboard

The entry point of this library is the module: Bopboard.

Library bopkit.bopkit

The entry point of this library is the module: Bopkit.

Library bopkit.cds-optimizer

The entry point of this library is the module: Bopkit_cds_optimizer.

Library bopkit.circuit

The entry point of this library is the module: Bopkit_circuit.

Library bopkit.command

The entry point of this library is the module: Bopkit_command.

Library bopkit.compiler

The entry point of this library is the module: Bopkit_compiler.

Library bopkit.counter

The entry point of this library is the module: Bopkit_counter.

Library bopkit.memory

The entry point of this library is the module: Bopkit_memory.

Library bopkit.pp

The entry point of this library is the module: Bopkit_pp.

Library bopkit.process

The entry point of this library is the module: Bopkit_process.

Library bopkit.process-command

The entry point of this library is the module: Bopkit_process_command.

Library bopkit.process-interpreter

The entry point of this library is the module: Bopkit_process_interpreter.

Library bopkit.process-pp

The entry point of this library is the module: Bopkit_process_pp.

Library bopkit.process-syntax

The entry point of this library is the module: Bopkit_process_syntax.

Library bopkit.pulse

The entry point of this library is the module: Bopkit_pulse.

Library bopkit.seven-segment-display

The entry point of this library is the module: Seven_segment_display.

Library bopkit.simulator

The entry point of this library is the module: Bopkit_simulator.

Library bopkit.sites

The entry point of this library is the module: Bopkit_sites.

Library bopkit.sleeper

The entry point of this library is the module: Bopkit_sleeper.

Library bopkit.subleq-command

The entry point of this library is the module: Subleq_command.

Library bopkit.subleq-debugger

The entry point of this library is the module: Subleq_debugger.

Library bopkit.subleq-generator

The entry point of this library is the module: Subleq_generator.

Library bopkit.subleq-simulator

The entry point of this library is the module: Subleq_simulator.

Library bopkit.syntax

The entry point of this library is the module: Bopkit_syntax.

Library bopkit.to_c

The entry point of this library is the module: Bopkit_to_c.

Library bopkit.topological-sort

The entry point of this library is the module: Bopkit_topological_sort.

Library bopkit.visa

The entry point of this library is the module: Visa.

Library bopkit.visa-assembler

The entry point of this library is the module: Visa_assembler.

Library bopkit.visa-command

The entry point of this library is the module: Visa_command.

Library bopkit.visa-debugger

The entry point of this library is the module: Visa_debugger.

Library bopkit.visa-dsl

The entry point of this library is the module: Visa_dsl.

Library bopkit.visa-pp

The entry point of this library is the module: Visa_pp.

Library bopkit.visa-simulator

The entry point of this library is the module: Visa_simulator.

Library bopkit.visa-syntax

The entry point of this library is the module: Visa_syntax.